fix toast icon class, rename
